+++ src/npc/plugin.rs
@@ -3,10 +3,10 @@ use bevy::prelude::*;
 
 use crate::{
     npc::{
-        components::NpcIdGenerator,
-        systems::{spawn_debug_npcs, update_schedule_state},
+        components::{NpcIdGenerator, ScheduleTicker},
+        systems::{spawn_debug_npcs, tick_schedule_state},
     },
-    world::{systems::spawn_world_environment, time::advance_world_clock},
+    world::systems::spawn_world_environment,
 };
 
 pub struct NpcPlugin;
@@ -14,7 +14,8 @@ pub struct NpcPlugin;
 impl Plugin for NpcPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<NpcIdGenerator>()
+            .init_resource::<ScheduleTicker>()
             .add_systems(Startup, spawn_debug_npcs.after(spawn_world_environment))
-            .add_systems(Update, update_schedule_state.after(advance_world_clock));
+            .add_systems(Update, tick_schedule_state);
     }
 }